requirements.txt
pyrogram
telegraph
tgcrypto
hachoir
pillow
asyncio
pySmartDL
